fix effect navigation iterator

+++ source/Types/include/Band.h
@@ -77,10 +77,10 @@ public:
   auto addEffect(haptics::types::Effect &newEffect) -> void;
   [[nodiscard]] auto isOverlapping(haptics::types::Effect &effect, const int start, const int stop)
       -> bool;
-  auto Evaluate(int position) -> double;
+  auto Evaluate(double position) -> double;
 
 private:
-  auto EvaluationSwitch(int position, haptics::types::Effect * effect) -> double;
+  auto EvaluationSwitch(double position, haptics::types::Effect *effect) -> double;
 
   BandType bandType = BandType::Wave;
   EncodingModality encodingModality = EncodingModality::Quantized;
